pyproject.toml
[project]
name = "dingtalk-approve"
version = "0.1.0"
description = "DingTalk approval event listener with AI spreadsheet integration"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dingtalk-stream>=0.24.3",
    "pyyaml>=6.0.0",
    "loguru>=0.7.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "watchfiles>=0.24.0",
]
